Don't shift CCD centers when we read them; instead use getPositionFromPixel() to transform to focal-plane coordinates.
Along with changes in afw that include the yaw in getPositionFromPixel(), this lets meas_mosaic meaningfully adjust chip rotations.
Hopefully this will also address problems in running meas_mosaic on cameras where getPixelSize() != 1, but that still needs testing.

It's possible I could have used cameraGeom routines (in particular, those that return an AffineTransform) to reimplement more of convertCoeffs, but it didn't appear to be worth the effort.