mosaicTask: remove checkInputs method
The checkInputs method assumes that the wcsDic iteration is in
the same order as the sourceSet and matchList, but it may not
be.  This can cause the wcsDic not to carry an entry for a
source, resulting in a NULL pointer dereference.  Instead,
have the getCatalog method remove empty frames; and generate
the wcsDic once the list of non-empty frames is known.