Limit unmatched sources to those within the tract.
meas_mosaic will solve the sky position of unmatched sources and
the number of those sources will dominate the matrix size to solve.
We will not care about outside of the tract. If we use unmatched
sources only within the tract, the number of sources and required
memory will be reduced. This feature can be controlled by config
parameter (clipSourcesOutsideTract). The default is True.