updateExposure: include Jacobian when correcting catalog
When we correct images, we have been applying both the Jacobian and the
polynomial corrections; this has been validated fairly extensively.
However, when we correct catalogs, we had been applying only the
polynomial correction (no Jacobian correction); apparently this code
had not been validated much at all, until now. The Jacobian correction
needs to be applied because the flat-field we have used is a measure of
surface brightness, so the constant(ish) brightness of the flat surface
is modified by the optics so that it varies over the focal plane.

Changed applyMosaicResultsCatalog so that it applies the Jacobian
correction, and changed getFluxFitParams so that it returns the
wcs so that the Jacobian correction can be calculated.