Accommodate running on HSC stack products
The HSC team is currently using a fork of the main LSST stack.  Many
developements have occurred on this fork for the purpose of HSC SSP
data release production runs.  What is currently the "HSC stack" will
eventually become obsolete, once all new development that happened
there is ported over to the LSST stack and validated.  For this
validation, it is useful to be able to directly LSST vs. HSC stack
runs when using the exact same inputs.  This is implemented here by
including accommodations required to parse HSC stack produced outputs,
namely a mapper between the different source schemas.  See the
readSrc() function in SourceReader class in MosaicTask.py along with
the config/hsc/Mosaic.py file in obs_subaru for an example of how to
apply this mapper.  The mapper itself is in config/hsc/srcSchemaMap.py
in obs_subaru.

Note that these inputs will already be in the coordinate system that
meas_mosaic expects (i.e. all CCDs have LLC pixel (0, 0) w.r.t the
focal plane coordinate system), so no accommodations need to be made
in that respect.