Accommodate running on HSC stack products
It is useful to compare the LSST vs. HSC stack runs when using
the exact same inputs.  This is implemented here by including
accommodations required to parse HSC stack produced outputs
(namely a mapper between the different source schemas).

Note that these inputs will already be in the coordinate system
that meas_mosaic expects, so not accommodations need to be made
in that respect.