Update applyMosaicResultsCatalog
Updates the application of the results to a catalog.  Since meas_mosaic's
solution is based on a coordinate system where a given detector's origin
(pixel 0, 0) is associated with its LLC w.r.t. to the focal plane.  LSST
always associates the detector origin with the electronics, thus if the
detector was rotated when inserted into the focal plane (nQuarter%4 != 0),
the centroids must be rotated to the meas_mosaic coordinate system before
applying the corrections.

Note that if working with catalogs produced by the HSC stack (hscRun is
not None), the pixels are already in the correct coordinate system, so
no rotations are necessary.