Rotate catalog centroids to meas_mosaic coords
If the detector has nQuarter%4 != 0 (i.e. it is rotated w.r.t the focal
plane coordinate system), the (x, y) pixel values of the centroid slot for
the source catalogs need to be rotated such that pixel (0, 0) corresponds
to the LLC (i.e. the coordinate system expected by meas_mosaic).