Tests for WCS rotation
Jim, the tests/xtestwcsRotation.py has known values for a given
source in a number of ccds based on the meas_mosaic results of
the rerun indicated in the rootDir:
rerun/lauren/LSST/DM-9862/cosmos_noJunk/orig/

If you just wanted to check that the RA and DEC get updated properly,
you can change the test line from:
result = applyMosaicResultsCatalog(dataRef, srcCat, True)
to
s.updateCoord(yourWcs())

and get rid of the other asserts.

If you want to go further, I've indicated where changes would need
to be made in mosaicTask.py and updateExposure.py if the WCS was
persisted in the LSST coords.  If you persist rotated WCS's that look
like they are meas_mosaic outputs, you can then indicate the rerun
that points to your new wcs_*.fits outputs for applyMosaicResultsCatalog
to you.  Obviously I can do this, but I'm not sure what would be easier
for you, so I'm giving you the option.

I couldn't think of a better way on this timescale to make this a
proper unittest (but would be willing to give it a try if what I've
provided won't work for you).