Rotate wcs to LSST coordinate system before persisting
Without this rotation, the wcs is output in the meas_mosaic coords (LLC
w.r.t. FP is 0,0) as opposed to the LSST coords (0,0 attached to the
electronics, so 0,0 is the associated with the LLC of ampifier 1).  This
makes use of the new rotateWcsPixelsBy90() function in meas_astrom, which
properly rotates a TAN-SIP wcs, to rotate the new wcs to the LSST
coordinate system prior to persisting.

Note that the fcr files are still being persisted in the HSC coordinate
system.  This is accounted for in updateExposure.py and is safe (as of
now) since no other code uses these files.