Remove maxtasksperchild=1 from pool initialization
This keyword had been added by commit 2ecc8fd but it has been
identified to cause indefinite hanging of the mosaic.py script
when run with numCoresForRead > 1---but only with the pybind11
version of the stack.  This is not a problem in the swig-wrapped
code (currently being used for the HSC pipeline stack).