Fix rotations when reading meas_mosaic solutions
Meas_mosaic writes out the wcs in the LSST coordinate frame,
but the photometric correction is still in the old HSC frame.
Therefore, when applying the meas_mosaic correction we need to
first rotate the wcs back to the HSC coordinate system to create the photometric
correction and then rotate the wcs badk to the LSST frame.