MosaicTask: set matplotlib backend
Setting the backend was removed because it was being done at the module
level. However, it is necessary for MosaicTask when we are using
diagnostics. Setting it at the top of mosaicTask.py, which isn't
imported by the module's __init__.py, so should not be visible at the
module level. This should give us the best of all worlds.