Fix correction of flux error fields
Prior to this, the loop into the error correction never triggered
as it was searching on the flux field name, but they are not the
same (the error key names have "Err", or some variation thereof
depending on the table version of the stack when the data were
processed appended to the flux name).  Instead, here we just loop
through and correct the entire list of errKeys returned by the
getFluxKeys funciton (where there is already a check on the
corresponding flux key before the error key is added to the list,
so an extra check here is redundant).