Read the fcr photoCalib directly
The fcr object is an empty Exposure, which now contains a PhotoCalib object,
persisted directly (instead of having Calib headers FLUXMAG0/FLUXMAG0ERR), so we
get the fcr_photoCalib, instead of reading the metadata. This should also maintain
backwards compatibility with old files.

Update the dataRef mock in test_fluxFitBoundedField to be able to return an
fcr_photoCalib (which is a valid composite object for a real butler).