Update shimCameraGeom and test for updated units
RFC-617 (and its implementation ticket DM-20548) updates the obs_subaru
camera geometry to use physical focal plane units of mm from the former
representation which used an effective scaling of "unity" (i.e.'
plateScale = 1 arcsec/mm, pixelSize = 1 mm).  This renders meas_mosaic
into a "broken" state as it relies deeply on the previous "unity" scaling.
While updating it to accommodate the unit update is not completely out of
the question, this would take significant effort and, given the adoption
by LSST and HSC of jointcal as the uber "calibration" algorithm, no
effort is currenlty being allocated to such an update.  As such, as of
the adoption of RFC-617, meas_mosaic is effectively in retirement.

As meas_mosaic is still part of lsst_distrib, the updates here are
simply the minimal set necessary to ensure the build succeeds with the
updated camera geometry.  The official retirement of meas_mosaic will
see it removed from lsst_distrib and likely moved from the lsst GitHub
organization to lsst-dm as a legacy repository.